--- app/graphql/mutations/add_opinion_mutation.rb
@@ -1,13 +1,0 @@
-# frozen_string_literal: true
-
-module Mutations
-  class AddOpinionMutation < BaseMutation
-    argument :input, Types::Custom::Inputs::Mutations::OpinionInput, required: true
-    type Types::Custom::Objects::Opinions::OpinionObject
-
-    def resolve(params)
-      super(params)
-      ::Opinions::AddOpinionService.call(params: @params)
-    end
-  end
-end
